pyproject.toml
[tool.poetry]
name = "crypto diary"
version = "1.0.0"
authors = ["Eugene Mozge <eumozge@gmail.com>"]
description = ""

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.2"
psycopg2-binary = "^2.9.6"
django-extensions = "^3.2.3"
django-environ = "^0.10.0"
django-cleanup = "^7.0.0"
django-imagekit = "^4.1.0"
pillow = "^9.5.0"
